Added a Lobby
The lobby is in game now and is accepting keyboard input. The game can
now start with three keyboard players, all using dfferent control
schemes. Still need to get gamepad lobby -> game stuff working, but the
framework for that is already in place. Next up? A more dynamic in-game
GUI.