Art is getting there!
So the game is functional in all aspects aside from gamepad controls.
Still need to get that done, but the framework is in place to make it
happen as soon as I have a gamepad to test with. In the meantime I've
begun art and sound should be on it's way soon enough.