Put items in the cart when thrown.
Items now go into the cart when they're thrown and get picked back up
when the cart is grabbed. This was done by adding some methods to
Item.cs and Player.cs as well as modifying the pickup and throw methods
of Player.cs. No new bugs have been spotted due to this change.