New fonts, pause button, button tweaks.
New fonts are in all of the menus, as well as some tweaked font sizes.
Each control scheme now has a pause button associated with it, but you
can't pause the game while you're dead. Need to fix that. I also tweaked
the lobby button behaviour. The start button is now selected as soon as
you have enough players joined, and the back button is reselected when
enough people leave.