New animations and tweaks
Added a player run animation without a cart. It MAY need to be tweaked to improve readability (it can be hard to tell what direction you're facing), but it's good enough for now. I also made animations run in reverse if the player is going in reverse. Lastly, I added another collider (a trigger) to the cart object in order to make it easier to pick up. You no longer need to ram into it while you grab, just be touching it.